refactor: simplify physical device selection logic
Moves the physical device selection logic into a new `PhysicalDevice` 
struct for better organization and encapsulation. Updates the main 
function to use the new struct, improving readability and maintainability. 
Removes unused code and hardcoded values, ensuring a cleaner codebase.
