+++ Cargo.toml
@@ -16,6 +16,3 @@ meshopt = "0.4.0"
 nalgebra-glm = { version = "0.19.0", features = ["convert-bytemuck"] }
 raw-window-handle = "0.6.2"
 tobj = { version = "4.0.2", features = ["ahash"] }
-
-[build-dependencies]
-anyhow = "1.0.92"
